requirements.txt
selenium == 4.5.0
pytest == 7.1.3
allure-pytest == 2.11.1